+++ server/endpoints/AdminChart.ts
@@ -1,5 +1,6 @@
+/* eslint-disable spaced-comment */
 import { body } from "express-validator";
-import { verifyToken } from "./utils";
+import { verifyToken, topSubjects as topSubjectsCB } from "./utils";
 import { Endpoint } from "../endpoints";
 import { Reviews, Classes, Subjects } from "../dbDefs";
 
@@ -7,7 +8,7 @@ import { Reviews, Classes, Subjects } from "../dbDefs";
 // howManyReviewsEachClass - done
 // totalReviews - done
 // getReviewsOverTimeTop15
-interface Token {
+export interface Token {
   token: string;
 }
 
@@ -16,12 +17,13 @@ interface GetReviewsOverTimeTop15Request {
   step: number;
   range: number;
 }
-
+/*
 export const getReviewsOverTimeTop15: Endpoint<GetReviewsOverTimeTop15Request> = {
   guard: [body("token").notEmpty().isAscii()],
   callback: async (request: GetReviewsOverTimeTop15Request) => {
+    const { token, step, range } = request;
     try {
-      const userIsAdmin = await Meteor.call<boolean>('tokenIsAdmin', token);
+      const userIsAdmin = await verifyToken(token);
       if (userIsAdmin) {
         const top15 = await Meteor.call<[string, number][]>('topSubjects');
         // contains cs, math, gov etc...
@@ -120,58 +122,11 @@ export const getReviewsOverTimeTop15: Endpoint<GetReviewsOverTimeTop15Request> =
       return null;
     }
   },
-};
+};*/
 
-export const topSubjects: Endpoint<unknown> = {
-  guard: [],
-  callback: async () => {
-    try {
-      // using the add-on library meteorhacks:aggregate, define pipeline aggregate functions
-      // to run complex queries
-      const pipeline = [
-        // consider only visible reviews
-        { $match: { visible: 1 } },
-        // group by class and get count of reviews
-        { $group: { _id: '$class', reviewCount: { $sum: 1 } } },
-        // sort by decending count
-        // {$sort: {"reviewCount": -1}},
-        // {$limit: 10}
-      ];
-      // reviewedSubjects is a dictionary-like object of subjects (key) and
-      // number of reviews (value) associated with that subject
-      const reviewedSubjects = new DefaultDict();
-      // run the query and return the class name and number of reviews written to it
-      const results = await Reviews.aggregate<{ reviewCount: number; _id: string }>(pipeline, () => { });
-      console.log(results);
-
-      await Promise.all(results.map(async (course) => {
-        const classObject = (await Classes.find({ _id: course._id }).exec())[0];
-        // classSubject is the string of the full subject of classObject
-        const subjectArr = await Subjects.find({ subShort: classObject.classSub }).exec();
-        if (subjectArr.length > 0) {
-          const classSubject = subjectArr[0].subFull;
-          // Adds the number of reviews to the ongoing count of reviews per subject
-          const curVal = reviewedSubjects.get(classSubject) || 0;
-          reviewedSubjects[classSubject] = curVal + course.reviewCount;
-        }
-      }));
-
-      // Creates a map of subjects (key) and total number of reviews (value)
-      const subjectsMap = new Map(Object.entries(reviewedSubjects).filter((x): x is [string, number] => typeof x[1] === "number"));
-      let subjectsAndReviewCountArray = Array.from(subjectsMap);
-      // Sorts array by number of reviews each topic has
-      subjectsAndReviewCountArray = subjectsAndReviewCountArray.sort((a, b) => (a[1] < b[1] ? 1 : a[1] > b[1] ? -1 : 0));
-
-      // Returns the top 15 most reviewed classes
-      return subjectsAndReviewCountArray.slice(0, 15);
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.log("Error: at 'topSubjects' method");
-      // eslint-disable-next-line no-console
-      console.log(error);
-      return null;
-    }
-  },
+export const topSubjects: Endpoint<Token> = {
+  guard: [body("token").notEmpty()],
+  callback: topSubjectsCB,
 };
 
 export const howManyEachClass: Endpoint<Token> = {
@@ -243,12 +198,14 @@ export const howManyReviewsEachClass: Endpoint<Token> = {
           },
         ];
         const results = await Reviews.aggregate<{ _id: string; total: number }>(pipeline, () => { });
-        results.map(async (data) => {
+
+        const ret = await Promise.all(results.map(async (data) => {
           const subNum = (await Classes.find({ _id: data._id }, { classSub: 1, classNum: 1 }).exec())[0];
           const id = `${subNum.classSub} ${subNum.classNum}`;
           return { _id: id, total: data.total };
-        });
-        return results;
+        }));
+
+        return ret;
       }
       return null;
     } catch (error) {
@@ -262,7 +219,7 @@ export const howManyReviewsEachClass: Endpoint<Token> = {
 };
 
 // Recreation of Python's defaultdict to be used in topSubjects method
-class DefaultDict<T = any> {
+export class DefaultDict<T = any> {
   [key: string]: T | Function;
 
   get(key: string): T | null {
